split param reading and time handling
