+++ app/src/main/res/layout/fragment_reading.xml
@@ -1,62 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimaryDark">
 
-    <LinearLayout
+
+    <TextView
+        android:id="@+id/tv_data_read"
+        style="@style/MediumTextViewStyleWhite"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:weightSum="4">
+        android:layout_weight="1"
+        android:gravity="bottom"
+        android:padding="@dimen/medium_text_padding"
+        android:scrollbars="vertical" />
 
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="3"
-            android:background="@color/light_gray"
-            android:gravity="center_horizontal"
-            android:orientation="vertical">
-
-            <TextView
-                android:id="@+id/tv_current_time"
-                style="@style/LargeTextViewStyleBoldWhite"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                tools:text="Device time: Sun-15:34:31" />
-
-            <DatePicker
-                android:id="@+id/date_picker"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:calendarTextColor="@color/white"
-                android:calendarViewShown="false"
-                android:datePickerMode="spinner"
-                android:spinnersShown="true" />
-
-            <Button
-                android:id="@+id/btn_time_pick"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/pick_time"
-                android:textAllCaps="false" />
-
-            <Button
-                android:id="@+id/btn_set_time"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/set_time" />
-        </LinearLayout>
-
-        <TextView
-            android:id="@+id/tv_data_read"
-            style="@style/MediumTextViewStyleWhite"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1"
-            android:gravity="bottom"
-            android:padding="@dimen/medium_text_padding"
-            android:scrollbars="vertical" />
-    </LinearLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
