+++ app/src/main/res/navigation/main_navigation_graph.xml
@@ -17,6 +17,14 @@
             app:exitAnim="@anim/fragment_fade_exit"
             app:popEnterAnim="@anim/fragment_fade_enter"
             app:popExitAnim="@anim/fragment_fade_exit" />
+
+        <action
+            android:id="@+id/navigateToDisplayTimeView"
+            app:destination="@id/displayTimeView"
+            app:enterAnim="@anim/fragment_fade_enter"
+            app:exitAnim="@anim/fragment_fade_exit"
+            app:popEnterAnim="@anim/fragment_fade_enter"
+            app:popExitAnim="@anim/fragment_fade_exit" />
         <action
             android:id="@+id/navigateToDisplayDataView"
             app:destination="@id/displayDataView"
@@ -38,14 +46,6 @@
             app:popEnterAnim="@anim/fragment_fade_enter"
             app:popExitAnim="@anim/fragment_fade_exit"
             app:popUpTo="@id/connectView" />
-        <action
-            android:id="@+id/navigateToDisplayTimeView"
-            app:destination="@id/displayTimeView"
-            app:enterAnim="@anim/fragment_fade_enter"
-            app:exitAnim="@anim/fragment_fade_exit"
-            app:popEnterAnim="@anim/fragment_fade_enter"
-            app:popExitAnim="@anim/fragment_fade_exit"
-            app:popUpTo="@id/connectView" />
     </fragment>
     <fragment
         android:id="@+id/displayDataView"
@@ -56,6 +56,6 @@
         android:id="@+id/displayTimeView"
         android:name="com.mtkreader.views.fragments.TimeView"
         android:label="DisplayTimeView"
-        tools:layout="@layout/fragment_display_time_data" />
+        tools:layout="@layout/fragment_time" />
 
 </navigation>
